refactor(chat): review cleanups for #665 — shared test fake, named factory type, throwing provider
- Consolidate the duplicated FakeConversationalAssistant into one shared
  fake in AnglesiteCoreTests.
- Name the factory closure type (FoundationModelAssistant.DesignInterviewModelFactory)
  and use it at every spelling; DesignInterviewTool.ModelProvider becomes
  throwing so a provider whose actor is gone fails loudly instead of
  silently fabricating an uncached interview.
- Read .site-config off the main actor in the chat factory closure.
